reviewer suggestions
- cleaned up import paths
- moved definitions from TabView.tsx to utils/types.ts
