+++ src/utils/types.ts
@@ -465,3 +465,20 @@ export interface ISubscriptionInstance {
     label?: string;
     [index: string]: any;
 }
+
+export const SUBSCRIPTION_VIEWTYPE_SELECTOR = "subscription-viewtype";
+
+export interface StoredViewPreferences {
+    viewType: string;
+    tabViewId: string;
+}
+
+export interface TabView {
+    id: string;
+    name: string;
+    href?: string;
+    content: React.ReactNode;
+    append?: string;
+    prepend?: string;
+    disabled: boolean;
+}
